Add dialog asking whether to delete incident transitions
Deleting a state that incident to any transitions may cause those transitions to be invalid. Hence, when deleting such a state, ask whether to remove the incident transitions as well. (Note that any deleted states are already automatically removed from transitions. The issue here is that transitions may be left with no current state or next states and therefore be invalid.)